#10 changed timers to use std::chrono. Fixed numerous type incompatabilities. Now compiles in windows

+++ ASMC_SRC/SRC/MemoryUtils.hpp
@@ -21,7 +21,7 @@
 
 #define MEM_ALIGNMENT 64
 
-void *ALIGNED_MALLOC(uint64 size);
+void *ALIGNED_MALLOC(size_t size);
 
 #ifdef USE_MKL_MALLOC
 #include <mkl.h>
